Fix Host Addition to DNS Dictionary
Problem Description: Prior to this update, the code encountered an error when attempting to add a new host record associated with two IPs to the DNS dictionary. The issue arose when the dictionary was already populated, and the program couldn't locate the host within it.

Implemented Solution: To address this problem, we have introduced a new function called addToDict, which is responsible for adding new host entries to the DNS dictionary every time a new host is added. Additionally, we have incorporated a variable named self.full_updates to maintain a list of domain names that have undergone full dictionary updates, as opposed to single entry additions. This new feature enables the program to determine when to search within the dictionary or when it is necessary to populate it.

These enhancements ensure that new host records are handled correctly, and their addition to the DNS dictionary is carried out seamlessly, regardless of the current state of the dictionary.