Added host record comment functionality
Added support for adding comments to host records.

If a new host record is being created and comments are added to the host record then they will get added to the host properties.

In the case where an existing host is being updated there are three options: add, append and replace.
Add: this will add a comment to the existing host record if a comment does not already exist on the record. If it does exist, the comment will be left as is.
Append: the comment will be added in a new line to the end of the existing comments. If there are no existing comments, the comment will just be added as per "add".
Replace: this will replace any existing comments with the new comment if applicable.

To add comments, add the comments to the 3rd row of the input CSV on a host entry with the comment action (add, append, replace) on the 4th row e.g.
host-record-1,10.10.10.10,This is an example comment,append